fix: split tests into pre-build and post-build suites
- Extract shared test helpers into test-utils.ts
- squircle-css.test.ts: CSS utility tests (depends on build)
- tw-plugin.test.ts: plugin tests (no build dependency)
- test:css task depends on build, test:plugin runs independently
- ready script simplified: fmt, lint, test (build runs via deps)
